+++ pyslvs_ui/graphics/color.py
@@ -7,6 +7,7 @@ __copyright__ = "Copyright (C) 2016-2020"
 __license__ = "AGPL"
 __email__ = "pyslvs@gmail.com"
 
+from typing import Tuple
 from qtpy.QtCore import QSize
 from qtpy.QtGui import QColor, QIcon, QPixmap
 from pyslvs import color_names, color_rgb
@@ -29,19 +30,14 @@ def color_icon(name: str, size: int = 20) -> QIcon:
     return QIcon(color_block)
 
 
-# Target path color: (road, dot, brush)
+# Target path color: (line, dot)
 _path_color = (
-    # Blue - Green
-    (QColor(69, 247, 232), QColor(3, 163, 120), QColor(74, 178, 176, 30)),
-    # Yellow - Green
-    (QColor(187, 221, 75), QColor(103, 124, 12), QColor(242, 242, 4, 30)),
-    # Red - Yellow
-    (QColor(252, 110, 27), QColor(237, 129, 66), QColor(242, 158, 109, 30)),
-    # Purple - Blue
-    (QColor(115, 0, 145), QColor(220, 104, 249), QColor(198, 137, 214, 30))
+    (color_qt('Dark-Gray'), color_qt('Black')),
+    (color_qt('Blue'), color_qt('Dark-Blue')),
+    (color_qt('Red'), color_qt('Dark-Red')),
 )
 
 
-def target_path_style(color_index: int) -> QColor:
+def target_path_style(color_index: int) -> Tuple[QColor, QColor]:
     """Get path colors."""
     return _path_color[color_index % len(_path_color)]